requirements.txt
requests==2.23.0
discord.py==1.3.3
python-decouple==3.3
pymongo==3.10.1
dnspython==1.16.0